Create Consolidated Frequent Words
Takes a dna_sequence string, iterated through each subsequence of length = kmer_length using a generator expression and populates a defaultdict with each kmer and its number of occurrences. A list is returned of all keys (kmers) whose value (occurrences) is equal to the maximum number of occurrences within the frequency dictionary.